Refactor Shadow Baking
Removed modal-related code in the shadow bake operator now that it only performs a shadow map bake.
Unfortunately, the current setup will suffer from the same problem as earlier when AO baking needs to be implemented (there's no way to know when it's done), so the AO-bake button might need to be split off too.

It might be possible to abuse Blender's logs. When a bake is finished, it seems to output "Baking map saved to internal image, save it externally or pack it" for each baked mesh separately.